refactor(frontend): extract chart sub-components and suppress bar chart legend
- Extract PieChartDisplay, BarChartDisplay, ChartTranscription into their own files
- Rename cardData prop to chart across chart components
- Use props: Readonly<Props> style instead of destructured signatures
- Hide the auto-generated "Serie 1" legend in BarChartDisplay via CSS wrapper
