+++ frontend/src/components/Analysis/AnalysisCard.tsx
@@ -1,25 +1,17 @@
 import { fr } from '@codegouvfr/react-dsfr';
-import Accordion from '@codegouvfr/react-dsfr/Accordion';
 import Alert from '@codegouvfr/react-dsfr/Alert';
-import { BarChart } from '@codegouvfr/react-dsfr/Chart/BarChart';
-import { PieChart } from '@codegouvfr/react-dsfr/Chart/PieChart';
 import Box from '@mui/material/Box';
-import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
 import Skeleton from '@mui/material/Skeleton';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 import { styled } from '@mui/material/styles';
-import type {
-  BarChartDataDTO,
-  DashboardCard,
-  PieChartDataDTO,
-  Resource
-} from '@zerologementvacant/models';
-import { Array, pipe } from 'effect';
+import type { DashboardCard, Resource } from '@zerologementvacant/models';
 import { match } from 'ts-pattern';
 
 import { useFindOneCardQuery } from '~/services/dashboard.service';
+
+import BarChartDisplay from './BarChartDisplay';
+import PieChartDisplay from './PieChartDisplay';
 
 interface Props {
   card: DashboardCard;
@@ -52,101 +44,9 @@ function formatValue(data: number, card: DashboardCard): string {
   }).format(data);
 }
 
-interface ChartTranscriptionProps {
-  labels: string[];
-  data: number[];
-  type: 'pie-chart' | 'bar-chart';
-}
+function AnalysisCard(props: Readonly<Props>) {
+  const { card, dashboardId } = props;
 
-function ChartTranscription({
-  labels,
-  data,
-  type
-}: Readonly<ChartTranscriptionProps>) {
-  const items = match(type)
-    .with('pie-chart', () => {
-      const total = pipe(
-        data,
-        Array.reduce(0, (acc, v) => acc + v)
-      );
-      if (total === 0) return [];
-      return pipe(
-        Array.zip(labels, data),
-        Array.map(([label, value]) => `${label} : ${Math.round((value / total) * 100)} %`)
-      );
-    })
-    .with('bar-chart', () =>
-      pipe(
-        Array.zip(labels, data),
-        Array.map(([label, value]) => `${label} : ${value}`)
-      )
-    )
-    .exhaustive();
-
-  return (
-    <Accordion label="Transcription">
-      <List>
-        {items.map((item, index) => (
-          <ListItem key={index}>{item}</ListItem>
-        ))}
-      </List>
-    </Accordion>
-  );
-}
-
-interface PieChartDisplayProps {
-  cardData: PieChartDataDTO;
-}
-
-function PieChartDisplay({ cardData }: Readonly<PieChartDisplayProps>) {
-  return (
-    <>
-      <PieChart
-        x={cardData.labels}
-        y={cardData.data}
-        name={cardData.labels}
-        color={[
-          'blue-france',
-          'blue-cumulus',
-          'blue-ecume',
-          'green-archipel',
-          'green-bourgeon',
-          'green-emeraude',
-          'green-menthe',
-          'green-tilleul-verveine'
-        ]}
-      />
-      <ChartTranscription
-        labels={cardData.labels}
-        data={cardData.data}
-        type="pie-chart"
-      />
-    </>
-  );
-}
-
-interface BarChartDisplayProps {
-  cardData: BarChartDataDTO;
-}
-
-function BarChartDisplay({ cardData }: Readonly<BarChartDisplayProps>) {
-  return (
-    <>
-      <BarChart
-        x={[cardData.labels]}
-        y={[cardData.data]}
-        horizontal={cardData.direction === 'horizontal'}
-      />
-      <ChartTranscription
-        labels={cardData.labels}
-        data={cardData.data}
-        type="bar-chart"
-      />
-    </>
-  );
-}
-
-function AnalysisCard({ card, dashboardId }: Readonly<Props>) {
   const { data, isLoading, isError } = useFindOneCardQuery({
     did: dashboardId,
     cid: card.id
@@ -190,10 +90,10 @@ function AnalysisCard({ card, dashboardId }: Readonly<Props>) {
 
       {match(data)
         .with({ type: 'pie-chart' }, (d) => (
-          <PieChartDisplay cardData={d} />
+          <PieChartDisplay chart={d} />
         ))
         .with({ type: 'bar-chart' }, (d) => (
-          <BarChartDisplay cardData={d} />
+          <BarChartDisplay chart={d} />
         ))
         .otherwise((d) => (
           <ShowcaseValue>{formatValue(d.data, card)}</ShowcaseValue>
